Implemented the following request from Orest (which was turned into a temporal task):
ID: 39
Name: Increase resolution reference values
Description:
Increase the resolution for displaying reference values in
the charting GUI. When the values being graphed are close
together (e.g., 1.2121, 1.2122, etc.), the resolution for
the reference values is too coarse, with the result that
not enough reference values are displayed.

Required, in BasicDrawer, adding some elements to the refvalue_specs array
to establish small ref values for small "yrange"s, adding an "x-adjustment"
for the 'ref_bounds_x_value' - to make more room for longer ref numbers.

Added a couple other small improvements.


In Utilities, needed to add a call to formatter.setMaximumFractionDigits(8);
to add the needed precision for smaller ref values.