Bug fix in draw_reference_values: When 'step' is very small, the line: 			start = (long) (Math.floor(ymin / step) * step + step);
causes the loop:

		for (y = start; y < ymax; y += step) {

to insert a very large number of values into 'y_values' because the floor
operation causes 'start' to be the integer value just below ymin; if
'step' is very small and 'ymin' is a large number of "step"s from start
the phenomenon of many values being inserted into 'y_values' occurs.
The bug is that this is an extreme waste of resources - CPU and memory -
and slows down the system noticeably.

The fix was to add the guard:

		if (step > 0.05) {

and the else clause:

	} else {
			start = ymin - step * 25;
		}

to avoid this situation when step is very small.