Removed most debug print statements.
The fix checked in last time caused a slight formatting problem:

In draw_reference_values the line:

			start = ymin - step * small_step_multiplier;

caused 'start' to often be ragged value, such as 8.1253001606

I added the line:

			start = Math.floor(start * 1 / step) * step;

To smooth it out based on the step value.  (Originally I did:

			start = Math.floor(start * 1000) * .001;

but basing it on 'step' correctly "smooths" the start value.)

Tested - looks good.