Added an 'append' feature to DataSet to accomodate the implementation of polling-based data retrieval, allowing newly retrieved data to be appended to an existing data set.
Implemented 'append' in DrawableDataSet.  Found that appending to a
double[] array was awkward, so I changed the type of the data in
DrawableDataSet to Vector (data, dates, and times).  This required fairly
extensive and time-consuming changes in other classes (mostly in graph/)
that depend on the type of the data set's data.  (Java's poor and
insconsistent design with respect to arrays and vectors made it much more
time consuming than it would otherwise be.)

Had to add an overloaded 'index_at_date' in Utilities to take a vector
argument.

Some cosmetic improvements (such as getting rid of leading _ in variable
names, rearranging sets of features in a more logical arrangement, etc.)
to DrawableDataSet and other classes.

Etc.

Everything compiles and a short test succeeded.  It could be that there is
a bug or two lurking that will be uncovered later.