Created a TimeDelimitedDataRequest class, descendant of TimerTask, that is responsible for reacting to a timer event by obtaining the needed parameters for a data request from its TimeDelimitedDataRequestClient (which in the OS GUI is an instance of the Chart class), making the request (using the client's data builder), appending to the appropriate associated data sets, and notifying the client that the update is complete.
TimeDelimitedDataRequestClient has a 'specification', which is a
TradableDataSpecification, holding the client's current tradable symbol,
period type, 'indicator_specifications' (IndicatorDataSpecification), etc.

Modified Chart to fulfill the roll of data-refresh client by
"implementing" TimeDelimitedDataRequestClient.

Created a AutoRefreshSetup class, which Chart uses to set up the timer
event scheduling needed to trigger the TimeDelimitedDataRequest.

And some other related changes.

The next step is probably to cleanup and complete the implementation of
AbstractDataSetBuilder.