Completed probably 95% or more of the auto-data-refresh feature implementation, except for testing and fixing resulting missing loose ends: Modified chart to use the tradable specification to hold its indicator information and to update its main and indicator data sets, etc. The specification is thus loaded with the appropriate data (Still need to test that the data is always updated correctly.) and the TimeDelimitedDataRequest object, when activated by the timer to handle a timer event will obtain the needed specs., make a query to the server for each data set in the specs that is not null (assuming that a null data set for an indicator means that the indicator is not selected to be graphed), and update the data set with the new data received from the server.
The above described behavior was minimally tested and appears to be
correct, but it was tested without a dynamically growing data supply on the
server side, so the data received would have been empty.  Need to test with
real data updates.  And test more cases, such as with multiple windows,
etc.

Several related changes.

Added -auto-refresh CL option, which turns on this feature.

Refined the Specification classes to fit better into the current scheme.

A bit of cleanup activity.

Added some debugging code, of course.

Etc.