Added the concept of indicator specifications being 'selected', meaning that the indicator is one the user has chosen to be displayed on the graph. Thus 'selected_indicators' from the TradableDataSpecification lists the indicators to be updated and displayed, and all other indicators contained by the TradableDataSpecification are to be ignored re. data updates and displaying.
The TimeDelimitedDataRequest object now calls 'selected_indicators' and
only retrieves and appends data (to the original data set) for "selected"
indicators.

Which indicators are selected is maintained in the Chart and
IndicatorListener classes.  (The IndicatorListener needs to be refactored
such that it delegates this duty to Chart.)

Also, TimeDelimitedDataRequest now does not bother to retrieve indicator
data sets if the main market data set request fails or is empty, cutting
down on unneeded network traffic.

Also, made a slight change to the semantics of 'append' in DataSet - the
argument is allowed to be null, resulting in a null op.

Cleaned up some obsoleted code in Chart.