Fixed these bugs, reported in a task 44 entry:
Uncovered a couple bugs causing a null pointer exception.  Fixed on by
adding a 'ready_for_request' query in the TimeDelimitedDataRequestClient.
The TimeDelimitedDataRequest now calls this function and if it's false
(e.g., if the initial data request gets an empty result because the
configured start date is > the latest date of the data from the server,
the data set will be empty and the last date time will be null and
'ready_for_request' is false) it does not proceed to prepare for
and make the data request.  (A new precondition for 'start_date' is
'ready_for_request'.)

Fixed another bug - with the auto refresh, sometimes the
IndicatorDrawer.x_values() function gives a null pointer exception because
'x_values' was not getting initialized.  The fix (tested successfully)
was to add an init routine for 'x_values' in MarketDrawer and call it
in the "public int[] x_values()" function if x_values is null, and to
call it in 'draw_tuples' in the Price- and Candle drawers.


Fixed this bug, reported in a task 44 entry:

"After these fixes, did a bit more debugging and discovered that,
apparently, the attachYdata routine of Axis ends up adding a data set to
its "dataset" list when it already has a data set, ending up with 2 data
sets.  I believe the data set should be replaced instead of added.
This may be one of the causes of the previously documented bugs."

Fixed by adding a 'resetDataSets' procedure that updates the data set state
(e.g., by calling 'range' on the data set), but does not call dataset.add.

This, unsurprisingly, fixes the problem of the data being drawn incorrectly
when auto-refresh is on and the user changes to a new tradable.

Fixed the
"Failed to read from server (java.net.SocketException: Socket is closed)"
problem that sometimes happen when changing tradables in the GUI conflicts
with the auto-refresh - Fixed by adding facilities to turn the refresh on
and off and turning it off before a complete data request and back on
after.

Fixed the problem in which current upper and lower indicators that were set
on startup from the settings file are not "selected" - They are not in the
list of indicators that need updating with fresh data from the server.
Fixed by, at the end of Chart.make_indicator_lists, calling the new
'select_indicators' procedure on tradable_specification with the current
upper and lower indicator lists, which by that time have been loaded from
the settings file.

Various other small fixes and improvements, including fixing some of the
graph classes - moving attributes to the bottom, public features to the
top, etc.