Created the BasicDataSet (which is to be used for updates - the dataset that will be created and appended to the main data set) by copying DrawableDataSet and removing unneeded features.  DrawableDataSet now inherits from BasicDataSet.  I believe most of the unneeded features have been stripped out (deleted or commented out) of BasicDataSet, but could have missed something.
DrawableDataSet needs to be edited to delete the duplicate features, those
that were not removed from BasicDataSet.

Parser needs to be changed to create basic data sets during updates.

Also, changed 'append' in DrawableDataSet to call range() and
*axis.resetRange() instead of resetDataSets.  I found this behavior in an
old append method in the original Brookshaw version and suspect it's still
correct for this class, but it needs to be tested.