Task 44 changes:
Fixed a bug in the parser classes - The parsing routines need to know
whether the parse if for an update or for a complete data set so that they
can create a BasicDataSet for an update and a DrawableDataSet for a
complete data set.

Chart and MA_ScrollPane: Fixed the problem with repainting the graph for a
data refresh update by adding a 'force_repaint_graphs' in MA_ScrollPane
and changing 'notify_of_update' in Chart to call that routine.  The
previous version ended up only updating the main graph if
'main_graph_changed' was true, which will often or usually be the case with
a data refresh.

Added synchronization to TimeDelimitedDataRequest.run to prevent one "timer
thread" from doing an update when the last one has not finished.

A partial fix to a null pointer exception encountered in
the 'append' routine of DrawableDataSet.  See the entry for task 44 with
"Date/time: 2004-10-28, 23:57" for more info. on the bug.

Found the cause of the "p.right() >= x_values.length" error in 'draw_day"
and 'draw_hour' of TimeDrawer: The 'x_values' routine in MarketDrawer will
only recreate 'x_values' if 'x_values' is null, so that if it was created
earlier with a smaller size, the out of bounds error is triggered.  The
fix:

Fix of bug in MarketDrawer - sometimes, in 'draw_day' and 'draw_hour'
of TimeDrawer, x_values would be accessed with a too-large index.  The fix
(in MarketDrawer):

< 		if (x_values == null) {
<			initialize_x_values();
---
> 		if (x_values == null || x_values.length != tuple_count()) {
>			initialize_x_values();


Added too many debugging print stmts.

Etc.