Added a new ChartableSpecification abstraction, allowing the possibility of mutliple tradables drawn in the same chart window.  Adjusted the Chart, TimeDelimitedDataRequest, and other affected classes to the use of this new abstraction.  Regression tests were successful.
Hid the "append to the data set" abstraction from TimeDelimitedDataRequest
somewhat by adding an 'append_data' routine to the TradableSpecification
and the IndicatorSpecification, so that data.append(...) does not need to
be called directly (allowing for an alternate design for "appending"
freshly arrived data to an existing data set).

Fixed a bug in 'draw_tuples' of TimeDrawer.