Implemented a fix for the following bug (reported as an entry for task 44):
"During auto-refresh requests when the configured start date for a period
type is not in the range of the server's data, the date/time for the
previous period type data is used instead of the configured start date
for that period type (because the 'latest_date_time' is null because
there is no data received from the server)."

From another log entry, a description of the fix:

"... [Fixed] by changing the server protocol to allow an empty start
date/time to signify that the start date specified for the requested
period type is to be used and, in the client code, allowing the start
date Calendar object to be null, in which case an empty string is sent to
the server; and changed the server to implement the changed protocol by
grabbing the start date for the requested period type from the 'session'
if the requested start date is empty."

The fix involved the following changes (so if a problem with the fix
occors, these would be the first places to look):

In "dispatch_market_data_request" of AbstractDataSetBuilder, remove the
guard after the call:

			Calendar date = data_parser.latest_date_time();

dealing with 'date' being null.  Thus, 'last_latest_date_time', which is
used by a client, is allowed to be null (which will occor when an empty
data set is received from the server).

In 'date_time_range' of NetworkProtocolUtilities, allowed the 'start'
argument to be null, which then results in 'start_string' (used to
construct the result) to be empty (i.e., the start date/time requested from
the server will be <null> (empty).

'date_time_range' is called by AbstractDataSetBuilder to build the data
request to send to the server.

In Chart:

In 'start_date', allow the result to be null and prevent calling
'protocol_util.one_second_later' with a null argument.

In 'ready_for_request', always return 'true' (implying that
ready_for_request is not needed, or some logic is missing).

(And some improvements in error handling related, I believe, to a different
problem.)

Did some testing to verify that the fix was correct - It appears to be, but
another bug fix is needed before I can test more thoroughly, as described
in the current last entry in the task 44 log:

"I believe the date/time-range bug ...
is now fixed.  Ran some more tests of the fix that looked promising,
but before I can verify it for sure I need to fix another bug that has
been noticable for a while: When hourly data is used and the configured
start date (as with the date/time-range bug) is after the last date of
the last current record, when the server's data "catches up" - data after
the start date is now available -, the GUI draws the data incorrectly -
it draws too many tuples (bars or candles) per hour.  ..."