Changed the scheme for testing whether an input file contains new data - Instead of checking the modification time and the size, just check the size.  This fixes a bug that shows up when using the auto-refresh facilities.  Here is a description of the bug and fix from a log entry for task 43:
> Noticed that the latest compile (and some previous ones, I think) are
> missing the last day of data.  I think the version before the above two
> changes was OK.  This needs to be fixed.

I discovered that the version before also had the bug and that it has
probably been present all along.  The fix:

In EXTENDED_FILE_BASED_TRADABLE_LIST, routine
'target_tradable_out_of_date_implementation', change:

				Result := status_work_file.date >
					current_file_status.last_modification_time and
					status_work_file.count > current_file_status.file_size

to:

				Result := status_work_file.count > current_file_status.file_size

In other words, don't check the "last modification time".  The bug would
manifest when the last modifcation time would be the same as the file's
current modification time, but the size would be different (perhaps because
the granularity of the OS's management of the modification time is not fine
enough to handle the frequency of changes being forced in the test).

Added some debugging code that will come out soon.