A bug fix and an improvement - from recent task entry:
In INTRADAY_TUPLE_DATA_SCANNER and MARKET_TUPLE_DATA_SCANNER - the
check_date_time routine.  In INTRADAY_TUPLE_DATA_SCANNER with assertions
on, the postcondition would produce assertions violations because it
was out of date with respect to logic used for auto-refresh updates.
Changed the postcondition to take the 'start_input' state into account:

< 			period_type_set_if_not_first_record: old last_time /= Void implies
< 				period_type /= Void

> 			period_type_set_if_start_input_and_not_first_record: start_input
> 				implies (old last_date_time /= Void implies period_type /= Void)

and changed the body accordingly.

I believe the only purpose of 'check_date_time' in
INTRADAY_TUPLE_DATA_SCANNER is to set the (intraday) period type if it
needs setting (which, I believe, will be used to set the target tradable's
period type).  When 'input_start' is true, the target tradable already has
a period type and this routine  (the one in INTRADAY_TUPLE_DATA_SCANNER)
is behaves as a null operation.

I also made a couple coding/structural improvements - notably, since
the 'last_time' attribute in INTRADAY_TUPLE_DATA_SCANNER is not really
needed (It's a duplicate of last_date_time in the parent class), I removed
it, made 'check_date_time' (in INTRADAY_TUPLE_DATA_SCANNER) into a hook
routine named 'auciliary_check_date_time', which is called in the parent,
MARKET_TUPLE_DATA_SCANNER at the right place for which last_date_time will
have the expected value.