The solution to the: "Client defect detected: Invalid request (Session may be stale.)" bug (that occurs when 2 or more windows are open, with auto-refresh on, and one of them gets closed) is simple:
It was caused by the window that closes itself (logs itself off from the
server) forgetting to turn off its auto-refresh timer.  Thus auto-refresh
requests for the window continued to be scheduled (with the now stale
session key).  Of course, the first time one of these requests goes to the
server, the stale key error comes back, upon which the process decides this
is a defect and should terminate.

The fix, of course, is - since Chart's turn_on_refresh routine (which
calls the auto-refresh handler's 'sechedule' routine) is called to start
up the timer, a call to turn_off_refresh (which calls the auto-refresh
handler's 'unsechedule' routine) was added to the sections of code
(I found two) that log out.

Tested briefly and appears to work, though I'll probably test a little more
to be sure.