Fixed the "no reference dates for EOD data" bug:
In 'temporal_drawer', since 'times' is now always initialized to an empty
list (needed by the auto-refresh functionality), the line:

 		if (times != null) result = time_drawer;

would resule, when EOD data is being processed, in the temporal drawer
being a time drawer instead of a date drawer.  And, of course, since a
TimeDrawer can't draw dates, no dates would be drawn for EOD data graphs.
The fix was simply to add the condition that 'times' must also not be empty
for 'temporal_drawer' to return the time drawer; otherwise (unless 'dates'
is null, which probably should not happen) it returns the date drawer.