Fixed a small bug in DrawableDataSet ctor: n > 0 -> n >= 0.
Found and fixed another a bug in 'append_data' in TradableSpecification -
occurs when composite data (e.g., hourly) is not in range (re. .maclientrc
settings) when the server's data file is small - and becomes in range
when the data files grows larger: occurs when 'current_data()' is empty -
don't remove the last record in this case.

Etc.