Some major restructuring/refactoring of 'send_data_request' to make it more manageable.
I'm going to change it further still - it will retrieve the main data set
and all needed indicator data sets and if there was no error, 'process'
these data sets - add them to the appropriate data structures, etc.  The
intent is to implement an "all or nothing" approach so that if an error,
such as a network time-out, occurs, no change is made to the existing data
structures.  This may fix the remaining "gui display messed up on time-out"
problems, but even if it doesn't, I believe it's a better, more robust
design.

I'm checking in the current refactored version here before making this
major change in case I need to revert to it, refer back to it, etc.