Added several debugging/print stmts for debugging the "messed-up-gui" problem.
Bug fix: Let 'end_date' be null instead of setting it to "now", which
allows the server's "now" to be used, which takes care of time-zone
differences between the client and server - the server time zone is what
counts.  The code that sets 'end_date' to:
	end_date = builder.last_latest_date_time();
ensures that the data queries all obtain data for the same time range.
(The old, commented-out, fix for the time-zone problem of adding 1 day to
"now" needs to be removed.)

Added 'ready_for_request' query to TimeDelimitedDataRequestClient and
implemented it in the descendant, Chart, to:
	return ! requesting_data.booleanValue() && ! is_exiting;
so that the TimeDelimitedDataRequest can abandon the request if the chart
is currently being updated (most likely because the user has requested a
change, such as changed tradables or period types).

Etc.