Some final cleanup with respect to the messed-up-gui fix before sending the current version to Orest.  Also reintroduced a couple changes, such as setting an unrealistic time-out value, to force time-out errors for a final test, which succeeded admirably - no detected problems.
Plan: Check this in with the test- changes.  Send the compiled files to
Orest for him to test.  Change it back to a reasonable time-out, etc.
Test once more, and then send Orest this version for some final testing.

As far as I can tell, the messed-up-gui bug is squished, but I'll not rule
it out for certain until Orest has had a chance to test it thoroughly.