Changed to better class name, created an EXPERIMENTAL_INPUT_DATA_CONNECTION, child of INPUT_DATA_CONNECTION as a "strategy" that simply sets the socket up to get data from an experimental test server (soon to be implemented).
SOCKET_TRADABLE_LIST now creates an EXPERIMENTAL_INPUT_DATA_CONNECTION and
uses it to connect to the data server.   It will probably be modified to
allow the 'connection' to be set by a client/builder class to allow dynamic
configuration of which INPUT_DATA_CONNECTION is used (depending on who the
data supplier is and what algorithms are needed to communicate with it).

And added other code related to implementing teh above classes.