Succeeded in moving the data request from the server/data-supplier to 'initialize_input_medium' (called by 'setup_input_medium') in SOCKET_TRADABLE_LIST.  Tested successfully with assertions.
Removed the postcondition in 'search_by_symbol' in TRADABLE_LIST that used
'item' and was causing the data to be retrieved within a postcondition,
thus preventing assertion checking for the data retrieval.  The change
exposed a couple strange assertion violations, including a "feature call on
Void target" where the feature was the 'count' routine.  This appeared to
happen during class invariant checking for 'execute' in TRADABLE_FACTORY
and the debugger gave no clue as to the path to the 'count' call or the
type of the object on which 'count' was being called.  I inserted an
"intermediary" routine to help figure out what was going on, but even this
did not help much.  But I was able to guess and verify that 'product' of
TRADABLE_FACTORY was void and someone (I still do not know who) was calling
'count' on it.  So I changed TRADABLE_FACTORY to create a dummy 'product'
in its make routine.

Several other changes, some related to the assertion-checking-related work
described above, some related to small and moderate improvements to the
socket-based input mechnaism.o

Enhanced the test supplier script to send back 5 lines of valid data.

Added intraday/set_intraday to TRADABLE_LIST (and removed from some
descendants) and added set_intraday call at appropriate place - it appears
appropriate for a TL to know whether it's intraday.

Etc.