Modified around 30 source files to adapt to changes between the estudio 5.1 compiler (the one I had been using) and the 5.6 compiler (which I'm currently trying out).  Most of the changes involve cleaning up incorrect constructs (such as "ensure" instead of "ensure then" and using an assertion of "i /= Void" when i is an INTEGER [thus expanded]).  A small minority of the changes, I believe, are not incorrect in the original, but the 5.6 compiler won't accept them (possibly because the language definition is a little bit in flux - re. ECMA standardization).
The result compiles fine (on the jupiter machine) with the 5.1 compiler,
and a brief test (the run_tests script and short usage test) succeeded.  It
currently does not compile with the 5.6 compiler because a few more
changes, most or all of which are not compatible with the 5.1 compiler,
need to be made before it will compile.

The current plan is to, after this checkin, create a branch whose first
revision consists of the remaining changes needed to get the system to
compile with the 5.6 compiler.