Merged estudio-5.6 branch change ( svn merge -r 4329:4330 file:///home/svn/finance/branches/estudio-5.6) onto the trunk.
[For convenience, here is the log entry for the change to the branch:

Bug fixes - bugs that apparently did not manifest under ES 5.1, but now
show up with 5.6: EXPONENTIAL_MOVING_AVERAGE: Change 'make' to initialize
attribute 'exp' at the beginning, instead of calling set_exponential at the
end.  This fulfills an assertion that would otherwise be violated in one of
the instructions that follow (might have been nrovf_set_n, but can't
remember for sure).  The 'sma_make' call was replaced with the instructions
that make up sma_make for the needed control the order of instructions.
Note that exp was set directly (and exp.set_n called) instead of calling
set_exponential because calling set_exponential would violate a constraint
caused by 'n' not yet being set.

And in 'make' in ONE_VARIABLE_FUNCTION, moved the make_output call from
the end to near the beginning. This satisfies an invariant condition (
output /= Void) that is unexpectedly violated while the creation routine
is executing because of a call of a feature of 'Current' that occurs as
a result of the initialization triggered by the creation routine.  Moving
make_output ensures that 'output /= Void' is true before this
initialization occurs.
]