Merged estudio-5.6 branch change ( svn merge -r 4335:4338 file:///home/svn/finance/branches/estudio-5.6 ) onto the trunk.
This was a change to function_editing_interface.e to implement a
work-around for a compiler bug that showed up in the 5.6 compiler.  The
change is not needed for the 5.1 compiler, but, although it uglifies the
code a bit, I merged the change into the trunk to keep the differences
between the trunk (which is currently maintained to compile with the ES 5.1
compiler) and the ES 5.6 branch.  Hopefully this "fix" can be removed in
the near future - when ISE/ESI releases a compiler without the bug.

I compiled and tested (with the 5.1 compiler) and verified that the
indicator editing functionality (affected by the fix) is the same - is
still correct.

(See the log entries for revs. 4336 - 4338 for function_editing_interface.e
for info. on the fix/work-around.)