Merged estudio-5.6 branch change ( svn merge -r 4342:4346 file:///home/svn/finance/branches/estudio-5.6 ) onto the trunk.
Some small restructuring to prepare for multi-threaded experiments in the
branch.  Incoporated into the trunk because it doesn't cause problems and
to keep the branch and the trunk from diverging significantly.

Also, included the addition of the gutentag*.ace files - ace files "fixed"
so that gutentag can create a tags file without errors.

For info. on the changes, read the log entires for -r 4343 .. 4346.

Compiled (finalized) and tested successfully.