mas.ecf, produced via 'ace2ecf mas.ace'
Added mas.ecf, which was created by running ace2ecf on mas.ace and then
running the eiffel compiler (via estudio) and fixing each error that it
reported.  estudio appears to now be happy with mas.ecf, but it might
still have problems that can't be detected yet, such as a missing
cluster/path spec (causing a needed class/file to not be found).