Finished changes needed to compile with ec 14.05.
Extensive changes to get MAS and the eiffel_library to compile with the
modern (14.05) eiffel compiler. The changes are pretty messy - e.g.,
commenting out necessary code that won't compile, and type changes and
"casts" that compile but may not be correct. This will all need to be
reviewed/fixed (with, possibly, some restructuring/refactoring) before the
system will both compile and run correctly. The system does compile, and
runs until an invariant violation occurs due to one of the type changes
mentioned above.