Made some progress on getting old MAS code working with ec 14.05
GLOBAL_APPLICATION: Changed error-recovery logic to, instead of "die"ing,
calling handle_exception to report the exception, with a stack trace,
(which is often caused by an assertion violation) so that I have more
information about why the program died.

To get rid of type-related problems - for example, the line:

    if attached {LIST [MARKET_FUNCTION]} Precursor as p then

failing to attach to p because a {LIST [MARKET_FUNCTION]} is not conformant
to a {LIST [TREE_NODE]}:  made MARKET_FUNCTION a descendant of
FUNCTION_PARAMETER; and made some related changes, such as changing most
instances of "LIST [TREE_NODE]" (which I changed from something else
earlier to get the system to compile) to "LIST [FUNCTION_PARAMETER]" (which
now "works" for all the MARKET_FUNCTION-related lists because of the new
inheritance relationship).  Also, needed to change this (in
MARKET_FUNCTION):

	descendants: LIST [MARKET_FUNCTION]
		do
			if attached {LIST [MARKET_FUNCTION]} Precursor as p then
				Result := p
			else
print ("DEBUG/14.05 conversion error: [p] not attached [MF.operators]%N")
			end
		end

to:

	descendants: LIST [MARKET_FUNCTION]
		local
			l: LIST [TREE_NODE]
			i: INTEGER
		do
			l := Precursor
			i := l.count
			from
				create {LINKED_LIST [MARKET_FUNCTION]} Result.make
				l.start
			until
				l.exhausted
			loop
				if attached {MARKET_FUNCTION} l.item as mf then
					Result.extend (mf)
				else
print ("DEBUG/14.05 conversion error: [p] not attached [MF.descendants]%N")
				end
				l.forth
			end
		end

That is, instead of casting Precursor to a LIST [MARKET_FUNCTION]}, which
failed, we cast each element of the list returned by Precursor to a
MARKET_FUNCTION (from a TREE_NODE) and insert those elements into a new
list (Result).  This runs successfully.

Added a bit of debugging code, temporary debug-related comments, and the
like.

Also, plugged in a temporary "sub-application" (files in src/server1/root)
as an experiment with covariance and how it works with the 14.05 compiler,
which I used as guidance in the MARKET_FUNCTION -> FUNCTION_PARAMETER
change.

Some other changes that I don't want to take the time to document.

The system now gets further before dying due to an assertion violation -
i.e., there's more work to do.