Bug fixed via change in eiffel_library and other minor changes.
Fixed a bug - actually in the eiffel_library, in COMMAND_SEQUENCE - that
was causing assertion violations/core-dumps when running the macl_scripts
script on a running executable to create new indicators used by the test
suite.  I believe the the bug was triggered by the copy_of_children attribute
not being set, causing access of a void value.  See more below in
description of the test/lib directory.  The macl_scripts script now runs
without segmentation or assertion violations and appears to succeed
(as far as I've investigated) completely.

STORABLE_LIST: call to ll_make in 'make' was commented out - Put it back
in, since there was no documentation on why it was commented out.  This
change doesn't appear to be causing incorrect behavior, but more testing is
needed.

Some changes to MAS_PRODUCT_INFO to give more information on the
date/status of the current executable (for debug/test purposes), although
the class is still in flux.

Some minor debugging-related changes.

mas-extended.ecf: Added GOBO time_zone cluster, although it may not be
needed.  And changed all backslashes into forward slashes.  This doesn't
matter to estudio/ec, but it "looks" better.

In the test/lib directory (src/server1/test/lib), (re-)added the
indicators_persist file for the old (<= 1.6.6) mas version as
indicators_persist.1.6.6  .  Also, changed the "regular" indicators_persist
file (the one for the current executable): updated with a structure change
(i.e., bug fix - removed copy_of_children attribute in COMMAND_SEQUENCE;
that is, restored it to a function [defined in TREE_NODE] - See above),
which, of course, makes the earlier version incompatible with the current
executable; and added some more indicators used by the tests (by running
macl_scripts with a running executable).