Eiffel Forum License -> GNU GPL v2
Changed the license from (for most of the files) Eiffel Forum License to
the GPL version 2.  Made the appropriate changes to each file - e.g., the
license section of the header comments for the Eiffel source files.  Also,
updated the copyright notice to 2014.  All code used in the MAS server
compiles after this change.  Other applications have not yet been compiled
with the estudio 14.05 compiler, so any problems in these other
applications introduced by this change will not be found until they are
compiled.