First commit for socket-enhancement: keep socket open
In preparation for keeping socket connections open after responding to a
client request, began implementing logic for flexibility in the
client/server communication protocol such that the server can be configured
to either close the socket connection after each response it sends (and,
via 'listen/accept', open a new connection when the client sends its next
request) or keep the socket open (until, according to the current plan, a
cache of connections grows beyond a limit).  Implementation of the latter
functionality is still in progress, but the former (which basically
implements the original behavior) is working, with some aspects duct-taped -
i.e., needs to be properly designed/refactored; working means it passes the
suite of tests in my (new) ruby-based client.  This new
design/implementation includes a new attribute, close_after_each_response,
in MAS_SOCKET_ACCEPTOR, which, when true, will cause the original behavior
(close after each send, re-open for new request, ...), as well as a new
class (in the eiffel_library), CONNECTED_SOCKET_POLL_COMMAND.  More
infrastructure needs to be put in place to support the new (keep connection
open) behavior.

Another added detail is an extension to the client/server protocol:
When the server will be closing the connection after sending a response, it
will send a different "OK" code compared to when it will not close the
connection - with a new constant named ok_will_close, so the client needs
to check for either the ok or the ok_will_close status and, if the status
is ok_will_close, get ready to open a new socket for its next request.
(This will, of course, be incompatible with old clients, such as the java
client - MA_Client.)  To complete this extension, alternate codes also need
to be implemented for the different error statuses - e.g.,
invalid_symbol/invalid_symbol_will_close, ...

Among other changes, the new implementation includes a redefinition of
'post_process', in MAS_SOCKET_ACCEPTOR, where (if close_after_each_response
is false) it creates a CONNECTED_SOCKET_POLL_COMMAND that holds the new
socket obtained via the 'accept' call.  This will be the socket that stays
open as long as the client is still around and as long as the (to-be-
implemented) cache has not filled.