Moved mas/finance-dependent file from eiffel_library to here.
Moved connected_socket_poll_command.e from the eiffel_library repository to
this one (by creating it as a new file here), since it depends on
CONNECTED_SOCKET_PROCESSOR in this repo. For the (short) history of the file
in eiffel_library, the former location (in eiffel_library) is
networking/connected_socket_poll_command.e.