Completion of socket-enhancement feature
Completed implementation of keep-socket-open behavior.  Completed logout
functionality when MAS_NO_CLOSE is in effect, with respect to the session
being logged out:

  - Close the associated socket.
  - Remove the associated CONNECTED_SOCKET_POLL_COMMAND from
    MEDIUM_POLLER's read list.
  - Mark the CONNECTED_SOCKET_POLL_COMMAND as expired so that it will be
    removed from the CONNECTION_CACHE the next time a new
    CONNECTED_SOCKET_POLL_COMMAND is added to the cache.

miscellaneous:

  - Made the CONNECTION_CACHE size configurable via environment variable
    MAS_CONN_CACHE_SIZE.
  - Got rid of the 'register_for_termination' call during initialization of
    the SOCKET_PROCESSOR objects (unless the new query
    'termination_required' is true).  This should allow the cached
    CONNECTED_SOCKET_POLL_COMMANDs to be garbage-collected when they are
    removed from the cache instead of waiting around to be "terminated".
  - 'post_process' in MAS_SOCKET_PROCESSOR only creates a
    CONNECTED_SOCKET_POLL_COMMAND and CONNECTED_SOCKET_PROCESSOR when
    is_non_persistent_connection is true, so command-line client
    connections will never use the CONNECTED_* objects.
  - CONNECTION_CACHE: 'add' now removes all items that have 'expired'
    (which state is set by the "logout" process) so that all
    unused/discarded CONNECTED_SOCKET_POLL_COMMANDs/SOCKETs are removed
    from the queue/cache before the new item is added.

Note: A bug appears to have been introduced (or it was already there from
previous work) in the implementation of communicating with the command-line
client: Sometimes the same command (issued from the client) needs to be
entered twice before the server executes it (where the first command is
responded to with an error message).  This does not happen all the time.  I
noticed it when I was running the java GUI client at the same time as macl
(MA_Client.class) for a test, so the bug might be triggered (possibly among
other scenarios) by running MA_Client and then macl.

eiffel_library:

Some small changes to support the mas/finance changes described above.