MARKET_PROCESSOR -> TRADABLE_PROCESSOR
Renamed MARKET_PROCESSOR to TRADABLE_PROCESSOR: The original name is not
the best fit, since this class models tradable items (i.e., an instance of
the class represents a tradable thing, such as a stock [e.g., IBM] or
commodity) rather than a market (where many tradable items can be bought or
sold) or markets.  The prefix "MARKET_" is used in the name of several
other classes, as well; these should probably also be changed in the same
way.

Two assertion-related bug fixes: check statement in the wrong place in
MAS_SESSION; and use of '=' (reference comparison) instead of '~'
postcondition of the make routine of SESSION_FUNCTION_PARAMETER.