MARKET -> TRADABLE, etc.
A rather large change to fulfill a seemingly simple description: Changed
most of the occurrences of MARKET to TRADABLE (e.g.,
MARKET_EVENT_REGISTRANT -> TRADABLE_EVENT_REGISTRANT) in class names (and
corresponding file-name change) and many occurrences of 'market' to
'tradable' in feature names.  The word "market" is not quite right to
describe, e.g., a stock, commodity, or option.  (See commit
09ac06d2b68cff3f122e0327d78d3734db279ba5 of
src/library/function_abstractions/tradable_processor.e for a more complete
reason for this name change.)  This change caused the indicators_persist
and generators_persist files in use at the time to become incompatible with
the executable.  Recreating the files exposed a bug that was
introduced during the OBJECT_INFO_REQUEST_CMD implementation that caused
creation of the *persist files from scripts to fail.  The fix involved the
'name' feature of command.e in the eiffel_library (which has been checked
in). Another, minor, bug (in command-line args parsing) was discovered in
the macl program - file src/clients/macl/support/macl_command_line.e -
now fixed.